+++ src/App.css
@@ -1,0 +1,3 @@
+.App {
+    height: 100vh;
+}
