--- src/components/MainLayout/MainLayout.module.css
@@ -1,3 +1,0 @@
-.main-layout {
-    height: 100vh;
-}
