+++ src/components/MainLayout/MainLayout.tsx
@@ -1,25 +1,11 @@
 import React from 'react';
-import {Layout, Menu} from "antd";
+import {Layout} from "antd";
 import AppRouter from "providers/router/AppRouter";
-import {Link} from "react-router-dom";
+import MainMenu from "components/MainMenu/MainMenu";
 
 const {Header, Sider, Content} = Layout;
 
 const MainLayout = () => {
-    const menuItems = [
-        {
-            key: 'users',
-            label: <Link to="/users">Пользователи</Link>,
-        },
-        {
-            key: 'roles',
-            label: <Link to="/roles">Роли</Link>,
-        },
-        {
-            key: 'profile',
-            label: <Link to="/profile">Мой профиль</Link>,
-        },
-    ];
 
     return (
         <Layout className='main-layout'>
@@ -27,7 +13,7 @@ const MainLayout = () => {
 
             <Layout>
                 <Sider>
-                    <Menu items={menuItems}/>
+                    <MainMenu />
                 </Sider>
                 <Content>
                     <AppRouter />
